requirements.txt
# pip install -r requirements.txt


aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
lxml==5.3.0
multidict==6.0.5
openai==0.28.0
packaging==24.1
pefile==2023.2.7
pydantic==2.8.2
pydantic_core==2.20.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-docx==1.1.2
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==72.2.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
